requirements.txt
certifi==2023.7.22
charset-normalizer==3.2.0
ci-info==0.3.0
click==8.1.7
configobj==5.0.8
configparser==6.0.0
etelemetry==0.3.0
filelock==3.12.2
future==0.18.3
httplib2==0.22.0
idna==3.4
isodate==0.6.1
looseversion==1.3.0
lxml==4.9.3
networkx==3.1
nibabel==5.1.0
nipype==1.8.6
numpy==1.25.2
packaging==23.1
pandas==2.0.3
pathlib==1.0.1
Pillow==10.0.0
prov==2.0.0
pydot==1.4.2
pyparsing==3.1.1
PyPDF2==3.0.1
python-dateutil==2.8.2
pytz==2023.3
pyxnat==1.6
rdflib==7.0.0
reportlab==4.0.4
requests==2.31.0
scipy==1.11.2
simplejson==3.19.1
six==1.16.0
traits==6.3.2
tzdata==2023.3
urllib3==2.0.4
